Initial Commit -RAnDome
RAnDome
By: TVeadedRobots
http://blog.tvheadedrobots.com/

RAnDome is a Processing Sketch that makes use of the monodic & oscP5
libraries as well as Processing's Perlin noise generator functions to
produce a generative, random animation using the monome's 8x8 matrix.
The reason I call the sketch generative is because the value of each
monomer row is then fed back into the noise generatorin orderr to
generate
the next rows value. In this way the sketch is self generating and very
random.

RAnDome by TVHeadedRobots is licensed under a Creative Commons
Attribution-ShareAlike 3.0 Unported License.